InfoWindow cleanup; SGE splitup
1. Start splitting SGE into more bite sized program units
2. Re-implement the logic for when to hide and show the infowindow
   and delegate it to the widgets. This also fixes many things.

Closes #4595
Addresses #4629
